+++ src/renderer/mod.rs
@@ -1,1 +1,2 @@
 pub mod window;
+pub mod mesh;
